+++ src/storage-preview/azext_storage_preview/_help.py
@@ -18,6 +18,43 @@ helps['storage account create'] = """
           text: az storage account create -n mystorageaccount -g MyResourceGroup -l westus --sku Standard_LRS
 """
 
+helps['storage account blob-inventory-policy'] = """
+type: group
+short-summary: Manage storage account Blob Inventory Policy.
+"""
+
+helps['storage account blob-inventory-policy create'] = """
+type: command
+short-summary: Create Blob Inventory Policy for storage account.
+examples:
+  - name: Create Blob Inventory Policy trough json file for storage account.
+    text: az storage account blob-inventory-policy create -g ResourceGroupName --account-name storageAccountName --policy @policy.json
+"""
+
+helps['storage account blob-inventory-policy show'] = """
+type: command
+short-summary: Show Blob Inventory Policy properties associated with the specified storage account.
+examples:
+  - name: Show Blob Inventory Policy properties associated with the specified storage account without prompt.
+    text: az storage account blob-inventory-policy show -g ResourceGroupName --account-name storageAccountName
+"""
+
+helps['storage account blob-inventory-policy update'] = """
+type: command
+short-summary: Update Blob Inventory Policy associated with the specified storage account.
+examples:
+  - name: Update Blob Inventory Policy associated with the specified storage account.
+    text: az storage account blob-inventory-policy update -g ResourceGroupName --account-name storageAccountName --set "policy.rules[0].name=newname"
+"""
+
+helps['storage account blob-inventory-policy delete'] = """
+type: command
+short-summary: Delete Blob Inventory Policy associated with the specified storage account.
+examples:
+  - name: Delete Blob Inventory Policy associated with the specified storage account without prompt.
+    text: az storage account blob-inventory-policy delete -g ResourceGroupName --account-name storageAccountName -y
+"""
+
 helps['storage account keys list'] = """
 type: command
 short-summary: List the access keys or Kerberos keys (if active directory enabled) for a storage account.