+++ src/storage-preview/azext_storage_preview/_params.py
@@ -3,19 +3,22 @@
 # Licensed under the MIT License. See License.txt in the project root for license information.
 # --------------------------------------------------------------------------------------------
 
-from azure.cli.core.commands.parameters import (get_enum_type, get_three_state_flag)
+from azure.cli.core.commands.parameters import (get_enum_type, get_three_state_flag, file_type)
+from azure.cli.core.local_context import LocalContextAttribute, LocalContextAction
 
 from ._validators import (get_datetime_type, validate_metadata,
                           validate_azcopy_upload_destination_url, validate_azcopy_download_source_url,
                           validate_azcopy_target_url, validate_included_datasets,
                           validate_blob_directory_download_source_url, validate_blob_directory_upload_destination_url,
-                          validate_storage_data_plane_list, ipv4_range_type,
+                          validate_storage_data_plane_list, process_resource_group, ipv4_range_type,
                           get_permission_help_string, get_permission_validator)
 from .profiles import CUSTOM_DATA_STORAGE_QUEUE
 
 
 def load_arguments(self, _):  # pylint: disable=too-many-locals, too-many-statements
+    from argcomplete.completers import FilesCompleter
     from knack.arguments import CLIArgumentType
+    from azure.cli.core.commands.parameters import get_resource_name_completion_list
 
     from .sdkutil import get_table_data_type
     from .completers import get_storage_name_completion_list, get_container_name_completions
@@ -25,6 +28,11 @@ def load_arguments(self, _):  # pylint: disable=too-many-locals, too-many-statem
     t_table_service = get_table_data_type(self.cli_ctx, 'table', 'TableService')
     t_queue_service = self.get_sdk('_queue_service_client#QueueServiceClient')
 
+    acct_name_type = CLIArgumentType(options_list=['--account-name', '-n'], help='The storage account name.',
+                                     id_part='name',
+                                     completer=get_resource_name_completion_list('Microsoft.Storage/storageAccounts'),
+                                     local_context_attribute=LocalContextAttribute(
+                                         name='storage_account_name', actions=[LocalContextAction.GET]))
     blob_name_type = CLIArgumentType(options_list=['--blob-name', '-b'], help='The blob name.',
                                      completer=get_storage_name_completion_list(t_base_blob_service, 'list_blobs',
                                                                                 parent='container_name'))
@@ -68,6 +76,52 @@ def load_arguments(self, _):  # pylint: disable=too-many-locals, too-many-statem
                    type=get_datetime_type(False))
         c.argument('if_match')
         c.argument('if_none_match')
+
+    with self.argument_context('storage account blob-inventory-policy') as c:
+        c.ignore('blob_inventory_policy_name')
+        c.argument('resource_group_name', required=False, validator=process_resource_group)
+        c.argument('account_name',
+                   help='The name of the storage account within the specified resource group. Storage account names '
+                        'must be between 3 and 24 characters in length and use numbers and lower-case letters only.')
+
+    with self.argument_context('storage account blob-inventory-policy create') as c:
+        # t_inventory_rule_type = self.get_models('InventoryRuleType', resource_type=CUSTOM_MGMT_PREVIEW_STORAGE)
+        c.argument('policy', type=file_type, completer=FilesCompleter(),
+                   help='The Storage Account Blob Inventory Policy, string in JSON format or json file path. '
+                        'See more details in: {https://review.docs.microsoft.com/en-us/azure/storage/blobs/'
+                        'blob-inventory?branch=pr-en-us-135665}.')
+        # c.argument('destination',
+        #            help='Container name where blob inventory files are stored. Must be pre-created.')
+        # c.argument('enabled', arg_type=get_three_state_flag(), help='Policy is enabled if set to true.')
+        # c.argument('type', arg_type=get_enum_type(t_inventory_rule_type), default='Inventory', required=False,
+        #            help='The valid value is Inventory. Possible values include: "Inventory".')
+        # c.argument('rule_name', arg_group='Blob Inventory Policy Rule',
+        #            help='A rule name can contain any combination of alpha numeric characters. Rule name is '
+        #            'case-sensitive. It must be unique within a policy.')
+        # c.argument('prefix_match', arg_group='Blob Inventory Policy Rule', nargs='+',
+        #            help='An array of strings for blob prefixes to be matched.')
+        # c.argument('blob_types', arg_group='Blob Inventory Policy Rule', nargs='+',
+        #            help='An array of predefined enum values. Valid values include blockBlob, appendBlob, pageBlob. '
+        #                 'Hns accounts does not support pageBlobs.')
+        # c.argument('include_blob_versions', arg_group='Blob Inventory Policy Rule', arg_type=get_three_state_flag(),
+        #            help='Include blob versions in blob inventory when value set to true.')
+        # c.argument('include_snapshots', arg_group='Blob Inventory Policy Rule', arg_type=get_three_state_flag(),
+        #            help='Include blob snapshots in blob inventory when value set to true.')
+
+    # with self.argument_context('storage account blob-inventory-policy rule') as c:
+    #     c.argument('destination', help='')
+    #     c.argument('enabled', help='')
+    #     c.argument('type', help='')
+
+    with self.argument_context('storage account network-rule') as c:
+        from ._validators import validate_subnet
+        c.argument('account_name', acct_name_type, id_part=None)
+        c.argument('ip_address', help='IPv4 address or CIDR range.')
+        c.argument('subnet', help='Name or ID of subnet. If name is supplied, `--vnet-name` must be supplied.')
+        c.argument('vnet_name', help='Name of a virtual network.', validator=validate_subnet)
+        c.argument('action', help='The action of virtual network rule.')
+        c.argument('resource_id', help='The resource id to add in network rule.')
+        c.argument('tenant_id', help='The tenant id to add in network rule.')
 
     with self.argument_context('storage blob service-properties update') as c:
         c.argument('delete_retention', arg_type=get_three_state_flag(), arg_group='Soft Delete',